requirements.txt
qdrant-client
python-dotenv
PyMuPDF
requests
openai
